Remove Sequence abstraction
