+++ src/main/java/sm808/Sequencer.java
@@ -1,24 +1,27 @@
 package sm808;
 
 import com.google.common.annotations.VisibleForTesting;
-import lombok.Getter;
+import com.google.common.base.Preconditions;
+import com.google.common.collect.HashMultimap;
+import com.google.common.collect.SetMultimap;
 import lombok.NonNull;
-import sm808.models.Sequence;
+import sm808.models.Event;
 import sm808.outputdevices.OutputDevice;
 
+import java.util.Set;
 import java.util.Timer;
 import java.util.TimerTask;
 
 public class Sequencer {
   @NonNull private final OutputDevice outputDevice;
-
-  /** The tempo, in BPM */
   private final int tempo;
-
   private final int subdivisions;
-  @Getter private Sequence sequence;
 
+  private int steps;
   private int currentStep = 0;
+  // A sequence is a map from a step's index to the set of events at that step.
+  private final SetMultimap<Integer, Event> sequence = HashMultimap.create();
+
   private Timer timer = new Timer();
 
   /**
@@ -37,7 +40,45 @@ public class Sequencer {
     this.outputDevice = outputDevice;
     this.tempo = tempo;
     this.subdivisions = subdivisions;
-    this.sequence = new Sequence(beatsPerSequence * this.subdivisions);
+    this.steps = beatsPerSequence * this.subdivisions;
+  }
+
+  /**
+   * Add the given event(s) to the sequence
+   * @param step The step to add the event at
+   * @param events One or more events to add
+   */
+  public void addEvents(final int step, @NonNull final Event... events) {
+    Preconditions.checkArgument(step < steps);
+    for (Event event : events) {
+      sequence.put(step, event);
+    }
+  }
+
+  /**
+   * Gets the events at a given step
+   * @param step The step
+   * @return The events for this step
+   */
+  public Set<Event> getEvents(int step) {
+    Preconditions.checkArgument(step < steps);
+    return sequence.get(step);
+  }
+
+  /**
+   * Clears all events from this sequencer's sequence
+   */
+  public void clear() {
+    sequence.clear();
+  }
+
+  /**
+   * Clears all events from the specified step
+   * @param step The step
+   */
+  public void clear(int step) {
+    Preconditions.checkArgument(step < steps);
+    sequence.get(step).clear();
   }
 
   /** Starts looping the sequence. */
@@ -61,10 +102,10 @@ public class Sequencer {
   /** Figures out which events need to be played right now and sends them to the output device. */
   @VisibleForTesting
   protected void click() {
-    outputDevice.play(sequence.getEvents(currentStep));
+    outputDevice.play(sequence.get(currentStep));
     currentStep++;
 
-    if (currentStep >= sequence.getSteps()) {
+    if (currentStep >= steps) {
       outputDevice.endBar();
       currentStep = 0;
     }
